feat: implement legacy doc URL decorator
The decorator prepends an HTML anchor to the description of each
operation with an HTML fragment from the legacy documentation.
This achieves link backward compatibility.

The legacy URL is stored on a custom property `x-legacy-doc-url` in the
operation object.

The PR populates the field in a few places so that it can be tested.